+++ src/css/_boomtable.scss
@@ -5,6 +5,12 @@
     .split {
         margin-right: 20px;
     }
+    .boom-btn {
+        padding: 10px;
+        margin-right: 10px;
+        min-width: 200px;
+        text-transform: lowercase;
+    }
     h6.text-header {
         padding: 0 0 10px 0;
         font-weight: bolder;
